Improve DB singleton implementation and rename namespace.
