+++ Controllers/Comparison/LecturerComparisonController.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Linq;
-using Advisor.Model;
+using Advisor.Models;
 using System.Collections.Generic;
 using Advisor.Service.Statistics;
