+++ Controllers/Course/CourseController.cs
@@ -1,4 +1,4 @@
-﻿using Advisor.Model;
+﻿using Advisor.Models;
 using System.Linq;
 using System.Collections.Generic;
 using Advisor.Service.Statistics;
