+++ Controllers/Faculty/FacultyController.cs
@@ -1,4 +1,4 @@
-﻿using Advisor.Model;
+﻿using Advisor.Models;
 using System;
 
 namespace Advisor.Controllers
