+++ Services/Statistics/StatisticCalculator.cs
@@ -1,4 +1,4 @@
-﻿using Advisor.Model;
+﻿using Advisor.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
