+++ Services/Validator/SignupDataValidator.cs
@@ -1,4 +1,4 @@
-﻿using Advisor.Model;
+﻿using Advisor.Models;
 using System.Text.RegularExpressions;
 
 namespace Advisor.Service.Validator
