+++ Tests/SignupDataValidatorTest.cs
@@ -1,4 +1,4 @@
-﻿using Advisor.Model;
+﻿using Advisor.Models;
 using Advisor.Service.Validator;
 using Xunit;
